backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
ultralytics==8.0.210
opencv-python==4.8.1.78
numpy==1.26.2
scipy==1.11.4
pydantic==2.5.2
pyserial==3.5